добавил транзакцию и разобрался с BulkInsert

+++ Application/Features/Categories/Commands/CreateCategory/CreateCategoryDTO.cs
@@ -16,6 +16,6 @@ namespace Application.Features.Categories.Commands.CreateCategory
         public string Title { get; set; }
         public string ImageUrl { get; set; }
         public string Description { get; set; }
-        public int? ParentId { get; set; }
+        public Int64? ParentId { get; set; }
     }
 }
