+++ Dmd.Services.Interfaces/ICategoryService.cs
@@ -10,7 +10,7 @@ namespace Dmd.Services.Interfaces
 {
     public interface ICategoryService
     {
-        public Task<Response<int>> Create(CreateCategoryDTO command);
+        public Task<Response<Int64>> Create(CreateCategoryDTO command);
         //public void Delete(int catId);
         //public void Save();
     }
